dom/tests/mochitest/whatwg/test_postMessage_basehref.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=414815
-->
<head>
  <title>postMessage's interaction with a &lt;base&gt; tag</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <base href="http://example.com/" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=414815">Mozilla Bug 414815</a>
<p id="display"></p>
<div id="content" style="display: none"></div>

<pre id="test">
<script class="testbody" type="application/javascript">
/** Test for Bug 414815 **/

SimpleTest.waitForExplicitFinish();

function receiveMessage(evt)
{
  is(evt.origin, "http://mochi.test:8888", "wrong sender");
  ok(evt.source === window, "wrong source");

  is(evt.data, "generate-event", "wrong data");
  is(evt.lastEventId, "", "wrong lastEventId");

  SimpleTest.finish();
}

window.addEventListener("message", receiveMessage);

function run()
{
  window.postMessage("generate-event", "http://mochi.test:8888");
}

window.addEventListener("load", run);
</script>
</pre>
</body>
</html>
